update(app): use components to named files for better IDE support
