+++ packages/app/src/components/CardScroller/index.tsx
@@ -1,63 +1,4 @@
-import {A as Link} from '@solidjs/router';
-import {type Component, For} from 'solid-js';
+import CardScroller from './CardScroller';
 
-import styles from './styles.module.scss';
-
-export type CardProps = {
-    href: string;
-    image: {
-        alt?: string;
-        'class'?: string;
-        src?: string;
-    };
-    title: string;
-};
-
-const Card: Component<CardProps> = props => {
-    return (
-        <Link
-            href={props.href}
-            class={styles.card}
-        >
-            <div class={styles.cardOverlay} />
-            <div class={styles.cardContent}>
-                <div class={styles.cardContentContainer}>
-                    <span class={styles.cardContentTitle}>
-                        {props.title}
-                    </span>
-                </div>
-            </div>
-            <div class={styles.cardBackground}>
-                <img
-                    class={props.image.class}
-                    src={props.image.src}
-                    alt={props.image.alt}
-                />
-            </div>
-        </Link>
-    );
-};
-
-const CardScroller: Component<{
-    items: CardProps[];
-    title: string;
-}> = props => {
-    return (
-        <div class={styles.cardScroller}>
-            <div class={styles.cardScrollerHeading}>
-                <h2>
-                    {props.title}
-                </h2>
-            </div>
-            <div class={styles.cardScrollerContent}>
-                <div class={styles.cardScrollerContentCards}>
-                    <For each={props.items}>
-                        {item => <Card {...item} />}
-                    </For>
-                </div>
-            </div>
-        </div>
-    );
-};
-
+export type {CardProps} from './Card';
 export default CardScroller;
